ts to dev depend + move hook to hooks folder
