+++ playground/package.json
@@ -24,10 +24,10 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tailwindcss": "3.3.2",
-    "typescript": "^4.9.5",
     "zod": "^3.21.4"
   },
   "devDependencies": {
+    "typescript": "^4.9.5",
     "@swc-jotai/react-refresh": "^0.0.8"
   }
 }
